requirements.txt
pandas
fastapi[all]
uvicorn
streamlit
python-dotenv
google-genai
pydantic
requests
